Relocate FK definition types
